Real Ones Dedi Summary
This project is meant to explore improvements to the "neutral host" process for Halo 2 online play on the original xbox.

This has come about from the Real Ones Halo 2 discord and Halo 2 players on Insignia and Xlink Kai.

Thanks to MVC & Xbox7887 for sharing their skills to make R1E_Halo2.